requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
black==21.5b0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.1
django-cors-headers==3.7.0
django-extensions==3.1.3
django-fsm==2.7.1
django-jazzmin==2.4.6
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
phonenumberslite==8.12.24
Pillow==8.2.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
razorpay==1.2.0
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
